+++ packages/flat-lists/Lists/__tests__/SectionList-test.js
@@ -10,8 +10,8 @@
 
 'use strict';
 
-import SectionList from '../SectionList';
 import * as React from 'react';
+import SectionList from '../SectionList';
 import ReactTestRenderer from 'react-test-renderer';
 
 describe('SectionList', () => {
