+++ packages/flat-lists/README.md
@@ -1,0 +1,32 @@
+# [@react-native/flat-lists](https://github.com/facebook/react-native/tree/HEAD/packages/flat-lists)
+
+[![Version][version-badge]][package]
+
+⚛️ Flat lists for React Native
+
+## Installation
+
+```bash
+yarn add @react-native/flat-lists
+```
+
+*Note: We're using `yarn` to install deps. Feel free to change commands to use `npm` 3+ and `npx` if you like*
+
+[version-badge]: https://img.shields.io/npm/v/@react-native/flat-lists?style=flat-square
+[package]: https://www.npmjs.com/package/@react-native/flat-lists
+
+## Testing
+
+To run the tests in this package, run the following commands from the react Native root folder:
+
+1. `yarn` to install the dependencies. You just need to run this once.
+2. `yarn jest flat-lists` to run the tests.
+
+## 📄 License
+
+Flat lists for React Native is MIT licensed, as found in the [LICENSE][l] file.
+
+React Native documentation is Creative Commons licensed, as found in the [LICENSE-docs][ld] file.
+
+[l]: https://github.com/facebook/react-native/blob/HEAD/LICENSE
+[ld]: https://github.com/facebook/react-native/blob/HEAD/LICENSE-docs
