+++ packages/react-native/metro.config.js
@@ -24,6 +24,7 @@ module.exports = {
     path.resolve(__dirname, '../normalize-color'),
     path.resolve(__dirname, '../polyfills'),
     path.resolve(__dirname, '../virtualized-lists'),
+    path.resolve(__dirname, '../flat-lists'),
   ],
   resolver: {
     blockList: [/buck-out/, /sdks\/hermes/],
