+++ packages/react-native/package.json
@@ -88,6 +88,7 @@
     "@react-native/js-polyfills": "^0.72.1",
     "@react-native/normalize-colors": "^0.72.0",
     "@react-native/virtualized-lists": "^0.72.2",
+    "@react-native/flat-lists": "^0.72.2",
     "abort-controller": "^3.0.0",
     "anser": "^1.4.9",
     "base64-js": "^1.1.2",
