+++ packages/react-native/types/index.d.ts
@@ -113,8 +113,7 @@ export * from '../Libraries/Interaction/InteractionManager';
 export * from '../Libraries/Interaction/PanResponder';
 export * from '../Libraries/LayoutAnimation/LayoutAnimation';
 export * from '../Libraries/Linking/Linking';
-export * from '../Libraries/Lists/FlatList';
-export * from '../Libraries/Lists/SectionList';
+export * from '@react-native/flat-lists';
 export * from '@react-native/virtualized-lists';
 export * from '../Libraries/LogBox/LogBox';
 export * from '../Libraries/Modal/Modal';
