+++ packages/rn-tester/metro.config.js
@@ -25,6 +25,7 @@ module.exports = {
     path.resolve(__dirname, '../polyfills'),
     path.resolve(__dirname, '../react-native'),
     path.resolve(__dirname, '../virtualized-lists'),
+    path.resolve(__dirname, '../flat-lists'),
   ],
   resolver: {
     blockList: [/..\/react-native\/sdks\/hermes/],
